+++ R/check-git.R
@@ -3,16 +3,20 @@
 #' This function performs Git checks checks prior to release. It is called
 #' automatically by [release()].
 #'
-#' @param pkg package description, can be path or package name.  See
-#'   [as.package()] for more information.
+#' @template devtools
 #' @keywords internal
 git_checks <- function(pkg = ".") {
   pkg <- as.package(pkg)
   cat_rule(paste0("Running Git checks for ", pkg$package))
 
+  git_report_branch(pkg)
   git_check_uncommitted(pkg)
   git_check_sync_status(pkg)
   cat_rule()
+}
+
+git_report_branch <- function(pkg) {
+  cat("Current branch:", git_branch(pkg$path), "\n")
 }
 
 git_check_uncommitted <- function(pkg) {