+++ R/zzz.R
@@ -48,13 +48,7 @@ NULL
 
 devtools_default_options <- list(
   devtools.path = "~/R-dev",
-  devtools.install.args = "",
-  devtools.name = "Your name goes here",
-  devtools.desc.author = 'person("First", "Last", email = "first.last@example.com", role = c("aut", "cre"))',
-  devtools.desc.license = "What license is it under?",
-  devtools.desc.suggests = NULL,
-  devtools.desc = list(),
-  devtools.revdep.libpath = file.path(tempdir(), "R-lib")
+  devtools.install.args = ""
 )
 
 .onLoad <- function(libname, pkgname) {